requirements.txt
readability-lxml
beautifulsoup4
Markdown
requests
openai
ujson